\name{add_s3_metadata}
\alias{add_s3_metadata}
\title{Add S3 metadata to a function}
\usage{
  add_s3_metadata(val, name, env)
}
\description{
  Add S3 metadata to a function
}
\seealso{
  Other Other s3: \code{\link{add_s3_metadata}}:
  \code{\link{is_s3_generic}}
}
% Generated by roxygen from [s3.r:67:81]
