% Built by roxygen3 0.1
\docType{class}
\name{AliasesTag-class}
\alias{AliasesTag-class}
\title{Add additional topic aliases.}
\description{
  Add additional aliases through which the user can find
  the documentation with \code{\link{help}}. The topic name
  is always included in the list of aliases.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @aliases space separated aliases
}
}
