% Built by roxygen3 0.1
\docType{class}
\name{AuthorTag-class}
\alias{AuthorTag-class}
\title{@author: the author of the documentation}
\description{
  A free text string describing the authors of the
  function.  This is typically only necessary if the author
  is not the same as the package author.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @author authors...
}
}
