\name{auto_imports}

\alias{auto_imports}

\title{Scans a function and determines what functions it uses from other packages}

\usage{
  auto_imports(f, name, imported = NULL, loaded = NULL)
}

\arguments{
  \item{f}{function to process}

  \item{imported}{list of functions that are already
  imported}
}

\description{
  Scans a function and determines what functions it uses
  from other packages.
}

\comment{Generated by roxygen from [R/roccer-ns-import.r]}

