% Built by roxygen3 0.1
\docType{class}
\name{Behaviour-class}
\alias{Behaviour-class}
\title{[DEV] An object representing processing behaviour.}
\description{
  This function is aimed primarily at developers extending
  this package. You shouldn't need to use it for everyday
  operation of the package.

  An object representing processing behaviour.
}
\keyword{classes}
