\docType{class}
\name{Block-class}

\alias{Block}
\alias{Block-class}

\title{Block class.}

\usage{
  Block(tags, object, srcref)
}

\description{
  The block class encapsulates the data about a single
  roxygen comment block, along with its location in the src
  code, and the object associated with the block.
}

\details{

}

% Generated by roxygen from [all-classes.r:31:42]

% Generated by roxygen from [roxy-block.r:8:24]
