% Built by roxygen3 0.1
\docType{class}
\name{Bundle-class}
\alias{Bundle-class}
\title{[DEV] An object representing a bundle of blocks in multiple files.}
\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

An object representing a bundle of blocks in multiple files.
}
\section{Class hierarchy for Bundle}{
Sub classes:
\itemize{
  \item \linkS4class{DirectoryBundle}
  \item \linkS4class{PackageBundle}
}

}

\section{Generics with methods for Bundle}{
\itemize{
  \item \code{\link{process}}
  \item \code{\link{rPath}} (2 methods defined for subclasses)
  \item \code{\link{writeDescription}} (1 methods defined for subclasses)
  \item \code{\link{writeNamespace}} (1 methods defined for subclasses)
  \item \code{\link{writeRd}} (1 methods defined for subclasses)
}

}

\keyword{classes}
