% Built by roxygen3 0.1
\docType{class}
\name{ClassHierarchyTag-class}
\alias{ClassHierarchyTag-class}
\title{@classHierarchy: describe class hierarchy for an S4 object.}
\description{
  This tag will add a new section listing all the super
  classes and all the subclasses of the object.  It will be
  compiled at the time that roxygen3 is run, and does not
  (yet) update dynamically as you load other packages and
  define new classes
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @classHierarchy
}
}
