% Built by roxygen3 0.1
\docType{class}
\name{CommentTag-class}
\alias{CommentTag-class}
\title{@comment: add comments to the Rd file.}
\description{
  @comment: add comments to the Rd file.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @comment This is a comment that will not be user visible.
}
}
