% Built by roxygen3 0.1
\docType{class}
\name{ConceptTag-class}
\alias{ConceptTag-class}
\title{Add a "concept" index entry.}
\description{
  These can be used to add additional entries that will be
  searched for in \code{\link{help.search}}.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @concept additional index entry
}
}
