% Built by roxygen3 0.1
\docType{class}
\name{DefaultExportTag-class}
\alias{DefaultExportTag-class}
\title{@defaultExports: stores information used by @exports}
\description{
@defaultExports: stores information used by @exports
}
\keyword{internal}
\section{Tag Usage}{
\preformatted{
#' # don't use - for internal use only
}
}
