% Built by roxygen3 0.1
\name{defaultTag}
\alias{defaultTag}
\title{[DEV] Generate default tag.}
\usage{
  defaultTag(tag, object)
}
\arguments{
  \item{tag}{class of tag}

  \item{object}{\linkS4class{Object} to generate defaults for.}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

Generate default tag.
}
\keyword{methods}
