% Built by roxygen3 0.1
\docType{class}
\name{DescriptionTag-class}
\alias{DescriptionTag-class}
\title{@description:: describe the function in one or two paragraphs.}
\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\linkS4class{IntroTag}} for more details.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @description Text goes here.
}
}
