\docType{class}
\name{DetailsTag-class}
\alias{DetailsTag-class}
\title{Override default details.}
\description{
  The topic title. By default this is taken from the third
  and subsequent paragraphs of the roxygen block. See
  \code{\link{tag_intro}} for more details.
}
\section{Tag Usage}{
  \code{ #' @details Text goes here. }
}

