% Built by roxygen3 0.1
\docType{class}
\name{DevTag-class}
\alias{DevTag-class}
\title{@dev: flag a function as for developers only.}
\description{
\code{@dev} flags a function as use primarily for developers building on
your package, rather than users. It adds \code{[DEV]} to the title and
a brief message to the description.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @dev
}
}
