% Built by roxygen3 0.1
\docType{class}
\name{EncodingTag-class}
\alias{EncodingTag-class}
\title{@encoding: specify non-standard encoding for Rd file}
\description{
  Generally you are safest to stick with ASCII encoding -
  if you need to provide documentation in another language,
  use this tag to specify how the text is encoded. The
  safest encodings to use are latin1, latin2 and UTF-8.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @encoding UTF-8
}
}
