% Built by roxygen3 0.1
\docType{class}
\name{FamilyTag-class}
\alias{FamilyTag-class}
\title{@family: associate the function with a family to automatically add
 cross-references.}
\description{
  All functions with the same family will have
  cross-references automatically added to all other
  functions in the family. '
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @family the name of the family'
}
}
