% Built by roxygen3 0.1
\docType{class}
\name{FormatTag-class}
\alias{FormatTag-class}
\title{@format: describe format of data.}
\description{
  If not specified, \code{str} will be used to generate a
  brief description.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @format A data.frame with 12 columns and 3 rows
}
}
