% Built by roxygen3 0.1
\docType{class}
\name{GenericMethodsTag-class}
\alias{GenericMethodsTag-class}
\title{@genericMethods: automatically list all methods of a generic}
\description{
@genericMethods: automatically list all methods of a generic
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @genericMethods
}
}
