\name{inherit_params}

\alias{inherit_params}

\title{Inherit parameters from another function.}

\usage{

}

\description{
  This tag will bring in all documentation for parameters
  that are undocumented in the current function, but
  documented in the source function. The source can be a
  function in the current package, \code{function}, or
  another package \code{package::function}.
}

\details{

}

% Generated by roxygen from [roccer-inherit-params.r:10:34]
