% Built by roxygen3 0.1
\docType{class}
\name{IntroTag-class}
\alias{IntroTag-class}
\title{Title, description and details.}
\description{
  Title, description and details.
}
\seealso{
  \code{\linkS4class{TitleTag}},
  \code{\linkS4class{DescriptionTag}},
  \code{\linkS4class{DetailsTag}} to set each component
  individually.
}
\section{Tag Usage}{
\code{
#' The first line becomes the topic title.
#' 
#' The next paragraph goes into the descrption section.  Any of these
#' sections can span multiple lines.
#' 
#' The third and subsequent paragraph go into the details.
}
}
