% Built by roxygen3 0.1
\docType{class}
\name{KeywordsTag-class}
\alias{KeywordsTag-class}
\title{@keywords: provide high-level keywords.}
\description{
@keywords: provide high-level keywords.
}
\details{
Roxygen3 will automatically add data, classes, and methods keywords for data,
objects (S4 and R5), and methods (S3 and S4).

Functions with the internal keyword will not be listed in the default
package index file.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @keywords keyword1 keyword2 keyword3
}
}
