\docType{class}
\name{NameTag-class}
\alias{NameTag-class}
\title{Override the default topic name.}
\description{
  By default, the topic name is derived from the object
  following the roxygen comments. This tag is rarely needed
  - if you are using it often it's probably an indication
  that you need to extend roxygen3 for you object type.
}
\section{Tag Usage}{
  \code{ #' @name name }
}

% Generated by roxygen from [tag-name.r:9:9]
