% Built by roxygen3 0.1
\docType{class}
\name{NoteTag-class}
\alias{NoteTag-class}
\title{@note: Add a note section.}
\description{
  Use this for special notes that the user should be aware
  of.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @note
#'  This text will be added to a note section.
#'
#'  You can use multiple paragraphs.
}
}
