% Built by roxygen3 0.1
\docType{class}
\name{Object-class}
\alias{Object-class}
\title{[DEV] The object being documented.}
\description{
  This function is aimed primarily at developers extending
  this package. You shouldn't need to use it for everyday
  operation of the package.

  The object being documented.
}
