\docType{class}
\name{PackageBundle-class}

\alias{PackageBundle}
\alias{PackageBundle-class}

\title{[DEV] A bundle of files in a package.}

\usage{
  PackageBundle(path, behaviour = default_behaviour())
}

\arguments{
  \item{path}{path of package root directory}

  \item{behaviour}{\code{\link{Behaviour}} object
  describing what tags, processors and writers to use}
}

\description{
  This function is useful only for developers

  The package class captures all the information about the
  package: its name, path, and all the \code{\link{Block}}s
  that it contains.
}

\details{

}

% Generated by roxygen from [all-classes.r:80:84]

% Generated by roxygen from [bundle-package.r:11:24]
