% Built by roxygen3 0.1
\docType{class}
\name{ParamTag-class}
\alias{ParamTag-class}
\title{@param: describe input parameters}
\arguments{
  \item{name}{Basic description. Multiple lines of text are
  fine, and are often useful if your parameter does a lot.}
}

\description{
  @param: describe input parameters
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' 
}
}
