\name{parse_block}
\alias{parse_block}
\title{Parse and execute a block of text in a package like environment.}
\usage{
  parse_block(text, behaviour = no_output())
}
\arguments{
  \item{text}{code to parse/execute}
}

\description{
  This is used cheifly for testing.
}
% Generated by roxygen from [test.r:25:42]
