\name{parse_directory}

\alias{parse_directory}

\title{[DEV]}

\usage{
  parse_directory(path, env = NULL)
}

\arguments{
  \item{path}{path to directory of R files to parse}

  \item{env}{environment that contains the results of
  evaluating the file. If \code{NULL}, will create a new
  environment with the global environment as a parent and
  will evaluate the code in that environment. This will be
  fine for simple, stand alone files, but you'll need to
  manage the environments yourself for more complicated
  packages.}
}

\description{
  This function is useful only for developers
}

\details{
  Parse directory of source files.
}

\comment{Generated by roxygen from [R/rocblock-read.r]}

