% Built by roxygen3 0.1
\name{process}
\alias{process}
\title{[DEV] Process bundles and blocks.}
\usage{
  process(input, ...)
}
\arguments{
  \item{input}{object to process}

  \item{...}{other arguments used by methods.}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

Process bundles and blocks.
}
\keyword{methods}
