\name{rd_out}

\alias{rd_out}

\title{[DEV]}

\description{
  This function is useful only for developers
}

\details{
  Output to Rd files in the man directory.

  This uses the \code{@rdname} tag to determine which file
  the output from each rocblock is sent to.

  Only one of \code{tag} and \code{out} can be supplied.
}

\comment{Generated by roxygen from [R/output-rd.r]}

