% Built by roxygen3 0.1
\docType{class}
\name{RdnameTag-class}
\alias{RdnameTag-class}
\title{@rdname: manually set the generated file name.}
\description{
  Since roxygen3 automatically translates non-alphanumeric
  components in the function name, you should only need to
  use this if you want to merge the documentation from
  multiple blocks into the same file.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @rdname name-of-file-without-extension.
}
}
