% Built by roxygen3 0.1
\docType{class}
\name{ReferencesTag-class}
\alias{ReferencesTag-class}
\title{' @references: add references to the literature.}
\description{
  ' @references: add references to the literature.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @references Text citation for an important paper.
}
}
