% Built by roxygen3 0.1
\docType{class}
\name{ReturnTag-class}
\alias{ReturnTag-class}
\title{@return: describe the output of the function}
\value{
What the function returns.  This can be a setnence of two
  or several paragraphs.
}
\description{
@return: describe the output of the function
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' 
}
}
