\name{roc_parser}

\alias{roc_parser}

\title{[DEV]}

\description{
  This function is useful only for developers
}

\details{
  Create a rocblock parser.

  The functions \code{tag}, \code{one} and \code{all} are
  called in that order, so if multiple are supplied, each
  can rely on having access to the results of the previous.
  See \code{\link{roc_name}} for an example of this.
}

\comment{Generated by roxygen from [R/roc-parse.r]}

