\name{rocblock_parser}

\alias{rocblock_parser}

\title{[DEV] A rocblock parser.}

\usage{
  rocblock_parser(process)
}

\arguments{
  \item{process}{A function that's given a list of all
  rocblocks and should return the data that the}
}

\description{
  This function is useful only for developers

  The rocblock parser is extremely general: the
  \code{process} function gets a list of all rocblocks and
  it returns the transformed output. This means it can
  perform actions that need a global perspective, as well
  as providing the ability to add and delete rocblocks.
}

% Generated by roxygen from [R/parser-rocblock.r]
