\name{roccer}

\alias{roccer}

\title{[DEV] Create a roccer object.}

\usage{
  roccer(name, parser = NULL, output = NULL)
}

\arguments{
  \item{name}{string giving the tag name. Do not include "}

  \item{parser}{parser to use to parse the tag. Currently
  roxygen3 only provides one parser:
  \code{\link{roc_parser}}.}

  \item{output}{output generated, like an object created by
  \code{\link{rd_out}} or \code{\link{namespace_out}}.}
}

\description{
  This function is useful only for developers

  The roccer object is a key component in roxygen3 - it
  defines the behaviour of a tag with a \code{parser} and a
  \code{output} write.
}

% Generated by roxygen from [R/roccer-.r]
