\docType{class}
\name{PackageBundle-class}

\alias{PackageBundle-class}

\title{PackageBundle class.}

\description{
  The package class captures all the information about the
  package: its name, path, and all the
  \code{\link{Block}}s that it contains.
}

\details{

}

% Generated by roxygen from [roxy-package.r:19:20]
