% Built by roxygen3 0.1
\docType{class}
\name{S3methodTag-class}
\alias{S3methodTag-class}
\title{Export S3 methods.}
\description{
  This method is only included for compatibility with
  previous version of roxygen.  Roxygen3 now automatically
  determines if a function is a S3 method and
  \code{@export} will produce the correct namespace
  directives.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @S3method generic class
#'  @S3method generic
#'  @S3method
}
}
