% Built by roxygen3 0.1
\docType{class}
\name{SectionTag-class}
\alias{SectionTag-class}
\title{@section: Add a new named section.}
\description{
  This tag allows you to add arbitrarily named new
  sections. The section name should not use any special
  formatting for latex compatibility.
}
\keyword{classes}
\section{Slots}{
\describe{
  \item{sections}{A named character vector. The names represent the section
 titles and the values the contents.}

}
}
\section{Tag Usage}{
\preformatted{
#' Section title:
#'  Section contents.  This can be multiple paragraphs.
}
}
