% Built by roxygen3 0.1
\docType{class}
\name{SeealsoTag-class}
\alias{SeealsoTag-class}
\title{@seealso: other places to look for documentation.}
\description{
  @seealso: other places to look for documentation.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @seealso Regular text, typically containing rd commands of the
#' form \\code\{\\link\{mean\}\}.
}
}
