% Built by roxygen3 0.1
\docType{class}
\name{SourceTag-class}
\alias{SourceTag-class}
\title{@source: the source of a dataset}
\description{
  @source: the source of a dataset
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @source A description of the source of the dataset, which might
#' include a \\url\{\} pointing to a source on the web.
}
}
