
\name{tag_aliases}

\alias{tag_aliases}

\title{Add additional topic aliases.}

\usage{
  #' @aliases space separated aliases
}

\description{
  Add additional aliases through which the user can find
  the documentation with \code{\link{help}}. The topic name
  is always included in the list of aliases.
}

% Generated by roxygen from [R/roccer-name-alias.r]
