
\name{tag_description}

\alias{tag_description}

\title{Override default description.}

\usage{
  #' @description Text goes here.
}

\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\link{tag__intro}} for more details.
}

% Generated by roxygen from [R/roccer-title-desc-details.r]
