\docType{class}
\name{ImportFromTag-class}

\alias{AutoImportsTag-class}
\alias{ImportClassesFromTag-class}
\alias{ImportFromTag-class}
\alias{ImportMethodsFromTag-class}
\alias{ImportTag-class}

\title{Namespace: tags for importing functions.}

\description{
  By and large, \code{@autoImports} should be the only
  imports tag that you need to use. It automatically
  generates the necessary \code{importFrom} statements to
  import all external functions used by this function.  See
  \code{\link{auto_imports}} for more implementation
  details.
}

\details{
  If there is a conflict, use \code{tag_importFrom} to
  resolve it. You can do \code{@importFrom base function} -
  this is not normally allowed in the \file{NAMESPACE}, but
  roxygen3 will simply ignore it, but still use it when
  resolving conflicts.

  You must have the packages declared in \code{DESCRIPTION}
  Imports.
}

\section{Tag Usage}{
  \code{ #' @importFrom package function1 function2 }
}

\section{Tag Usage}{
  \code{ #' @autoImports }
}

\section{Tag Usage}{
  \code{ #' @import package1 package2 package3 }
}

\section{Tag Usage}{
  \code{ #' @importClassesFrom package fun1 fun2 }
}

\section{Tag Usage}{
  \code{ #' @importMethodsFrom package fun1 fun2 }
}

% Generated by roxygen from [tag-namespace-import.r:19:20]

% Generated by roxygen from [tag-namespace-import.r:47:47]

% Generated by roxygen from [tag-namespace-import.r:72:72]

% Generated by roxygen from [tag-namespace-import.r:83:83]

% Generated by roxygen from [tag-namespace-import.r:94:94]
