
\name{tag_include}

\alias{tag_include}

\title{Control order of collation.}

\description{
  Control order of collation.
}

\details{
  The collation order is only modified if it is different
  from alphabetical order (in the C locale). That is R's
  default ordering.

  @usage @include file-1.r file-2.r
}

% Generated by roxygen from [R/roccer-include.r]
