
\name{tag_inheritParams}

\alias{tag_inheritParams}

\title{Inherit parameters from another function.}

\usage{
  #' @inheritParams source_function
}

\description{
  This tag will bring in all documentation for parameters
  that are undocumented in the current function, but
  documented in the source function. The source can be a
  function in the current package, \code{function}, or
  another package \code{package::function}.
}

% Generated by roxygen from [R/roccer-inherit-params.r]
