
\name{tag_name}

\alias{tag_name}

\title{Override the default topic name.}

\usage{
  #' @name name
}

\description{
  By default, the topic name is derived from the object
  following the roxygen comments. This tag is rarely needed
  - if you are using it often it's probably an indication
  that you need to extend roxygen3 for you object type.
}

% Generated by roxygen from [R/roccer-name-alias.r]
