
\name{tag_ns_auto_import}

\alias{tag_ns_auto_import}

\title{Namespace: automatically generate import statements.}

\usage{
  #' @auto_import
}

\description{
  Automatically generates the necessary \code{importFrom}
  statements to import all external functions used by this
  function.  See \code{\link{auto_imports}} for more
  implementation details.
}

\details{
  If there is a conflict, use \code{tag_importFrom} to
  resolve it.

  You must have the packages declared in \code{DESCRIPTION}
  Imports.
}

% Generated by roxygen from [R/roccer-ns-import.r]
