
\name{tag_ns_import_from}

\alias{tag_ns_import_from}

\title{Namespace: import function from package.}

\usage{
  #' @importFrom package function1 function2
}

\description{
  Produces multiple \code{importFrom(package, function)}
  directives to import selected functions from a package.
  This is mainly useful for resolving conflicts generated
  by \code{\link{tag_auto_import}}.
}

\details{
  You can do \code{@importFrom base function} - this is not
  normally allowed in the \file{NAMESPACE}, but roxygen3
  will simply ignore it, but still use it when resolving
  conflicts.
}

% Generated by roxygen from [R/roccer-ns-import.r]
