
\name{tag_roc_doctype}

\alias{tag_roc_doctype}

\title{Set object documentation type.}

\usage{
  #' @docType date @docType package @docType custom doctype
}

\description{
  Set object documentation type.
}

\details{
  You can use any doctype, but it will only be included in
  the Rd file if it is one of the standard R doctypes:
  data, package, methods and class.
}

% Generated by roxygen from [R/roccer-doctype.r]
