\docType{roccer}

\name{@title}

\alias{@title}

\title{Override default topic title.}

\usage{
  #' @title Topic title
}

\description{
  The topic title. By default this is taken from the first
  paragraph of the roxygen block. See
  \code{\link{roc_intro}} for more details.
}

% Generated by roxygen from [R/roccer-title-desc-details.r]
