\docType{class}
\name{TagAliases}

\alias{TagAliases}

\title{Add additional topic aliases.}

\description{
  Add additional aliases through which the user can find
  the documentation with \code{\link{help}}. The topic name
  is always included in the list of aliases.
}

\details{

}

\section{Tag Usage}{
  \code{ #' @aliases space separated aliases }
}

% Generated by roxygen from [tag-aliases.r:9:9]
