\docType{class}
\name{TagAliases}

\alias{}
\alias{TagAliases}

\title{Add additional topic aliases.}

\usage{

}

\description{
  Add additional aliases through which the user can find
  the documentation with \code{\link{help}}. The topic name
  is always included in the list of aliases.
}

\details{

}

