\docType{class}
\name{TagComment}

\alias{TagComment}

\title{Add comments to the Rd file.}

\description{
  Add comments to the Rd file.
}

\details{

}

\section{Tag Usage}{
  \code{ #' @comment This is a comment that will not be
  user visible. }
}

% Generated by roxygen from [tag-comments.r:5:5]
