\docType{class}
\name{TagDescription-class}

\alias{TagDescription-class}

\title{Override default description.}

\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\link{tag_intro}} for more details.
}

\details{

}

\section{Tag Usage}{
  \code{ #' @description Text goes here. }
}

% Generated by roxygen from [tag-description.r:7:7]
