\docType{class}
\name{TagDescription}

\alias{TagDescription}

\title{Override default description.}

\usage{
  @description Text goes here.
}

\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\link{tag_intro}} for more details.
}

\details{

}

% Generated by roxygen from [tag-description.r:7:7]
