\docType{class}
\name{TagDetails}

\alias{TagDetails}

\title{Override default details.}

\description{
  The topic title. By default this is taken from the third
  and subsequent paragraphs of the roxygen block. See
  \code{\link{tag_intro}} for more details.
}

\details{

}

\section{Tag Usage}{
  \code{ #' @details Text goes here. }
}

% Generated by roxygen from [tag-details.r:8:8]
