\docType{class}
\name{TagName}

\alias{}
\alias{TagName}

\title{Override the default topic name.}

\usage{

}

\description{
  By default, the topic name is derived from the object
  following the roxygen comments. This tag is rarely needed
  - if you are using it often it's probably an indication
  that you need to extend roxygen3 for you object type.
}

\details{

}

% Generated by roxygen from [tag-name.r:9:9]
