\docType{class}
\name{TagTitle}

\alias{TagTitle}

\title{Override default topic title.}

\usage{
  @title Topic title
}

\description{
  The topic title. By default this is taken from the first
  paragraph of the roxygen block. See
  \code{\link{tag_intro}} for more details.
}

\details{

}

% Generated by roxygen from [tag-title.r:7:7]
