\name{usage}

\alias{usage}

\title{[DEV] Generate usage statement for an R object.}

\usage{
  usage(obj, name)
}

\arguments{
  \item{obj}{object to generate usage for}

  \item{name}{name of object (if not computable from
  \code{obj})}
}

\description{
  This function is useful only for developers

  Generate usage statement for an R object.
}

% Generated by roxygen from [R/usage.r]
