% Built by roxygen3 0.1
\docType{class}
\name{UsageTag-class}
\alias{UsageTag-class}
\title{@usage: describe the usage of a function.}
\description{
Roxygen3 will automatically generate the usage string for you - if you need
specify this yourself (unless you particularly want to control the formatting
it's probably an indiciation of roxygen bug and should be reported).
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @usage f(a, b, c) # You can also comments
}
}
