\docType{methods}
\name{writeNamespace}

\alias{writeNamespace}

\title{[DEV] An output generator for the \file{NAMESPACE} file.}

\usage{
  \S4method{writeNamespace}{RoxyPackage}(object)
}

\arguments{
  \item{tag}{function that processes a single tag. It
  should return a character vector of lines to be included
  in the \file{NAMESPACE}.  Duplicates will be
  automatically removed.}

  \item{name}{input tag name, usually set by
  \code{\link{roccer}}.}
}

\description{
  This function is useful only for developers

  An output generator for the \file{NAMESPACE} file.
}

\details{

}

% Generated by roxygen from [write-namespace.r:9:11]
