% Built by roxygen3 0.1
\name{add_s3_metadata}
\alias{add_s3_metadata}
\title{Add S3 metadata to a function}
\usage{
  add_s3_metadata(val, name, env)
}
\arguments{
  \item{val}{funtion object}

  \item{name}{name of the function}

  \item{env}{environment in which to search}
}

\description{
Add S3 metadata to a function
}
\seealso{
Other s3: \code{\link{is_s3_generic}}
}
\keyword{interal}
