% Built by roxygen3 0.1
\docType{class}
\name{AuthorTag-class}
\alias{AuthorTag-class}
\title{Provide the object's author.}
\description{
  A free text string describing the authors of the
  function.  This is typically only necessary if the author
  is not the same as the package author.
}
\section{Tag Usage}{
  \code{ #' @author authors... }
}

