\docType{class}
\name{Block-class}

\alias{Block}
\alias{Block-class}

\title{Block class.}

\usage{
  Block(tags = list(), object = new("NullObject"),
    srcref = new("NullSrcref"))
}

\description{
  The block class encapsulates the data about a single
  roxygen comment block, along with its location in the src
  code, and the object associated with the block.
}

% Generated by roxygen from [all-classes.r:38:49]

% Generated by roxygen from [block.r:8:24]
