% Built by roxygen3 0.1
\docType{class}
\name{Block-class}
\alias{Block}
\alias{Block-class}
\title{Block class.}
\usage{
  Block(tags = list(), object = new("NullObject"),
    srcref = new("NullSrcref"), default_tags = base_tags())
}
\description{
  The block class encapsulates the data about a single
  roxygen comment block, along with its location in the src
  code, and the object associated with the block.
}
