% Built by roxygen3 0.1
\docType{class}
\name{Bundle-class}
\alias{Bundle-class}
\title{[DEV] An object representing a bundle of blocks in multiple files.}
\description{
  This function is aimed primarily at developers extending
  this package. You shouldn't need to use it for everyday
  operation of the package.

  An object representing a bundle of blocks in multiple
  files.
}
