% Built by roxygen3 0.1
\docType{class}
\name{Bundle-class}
\alias{Bundle-class}
\title{[DEV] An object representing a bundle of blocks in multiple files.}
\description{
  This function is aimed primarily at developers extending
  this package. You shouldn't need to use it for everyday
  operation of the package.

  An object representing a bundle of blocks in multiple
  files.
}
\section{Class hierarchy for Bundle}{
  Sub classes: \itemize{ \item
  \linkS4class{DirectoryBundle} (S4) \item
  \linkS4class{PackageBundle} (S4) }
}

\section{Generics with methods for Bundle}{
  \itemize{ \item \code{\link[=process]{process}} \item
  \code{\link[=rPath]{rPath}} \item
  \code{\link[=writeDescription]{writeDescription}} \item
  \code{\link[=writeNamespace]{writeNamespace}} \item
  \code{\link[=writeRd]{writeRd}} }
}

\keyword{classes}
