% Built by roxygen3 0.1
\docType{class}
\name{ClassMethodsTag-class}
\alias{ClassMethodsTag-class}
\title{@classMethods: automatically list all methods of an object}
\description{
  @classMethods: automatically list all methods of an
  object
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @classMethods # exclude inherited methods
#'  @classMethods inherited # include inherited methods
}
}
