\docType{class}
\name{DebugTag-class}
\alias{DebugTag-class}
\title{Automatically add the path as an Rd comment.}
\description{
  This is useful for debugging Rd files, if you're not sure
  where the components come from. It is not enabled by
  default - you need to use \code{defaultBehaviour(debug =
  TRUE)} to turn it on.
}
\section{Tag Usage}{
  \code{ #' None: this is added automatically. }
}

