\name{default_behaviour}

\alias{default_behaviour}

\title{[DEV] Encapuslate default behaviour.}

\usage{
  default_behaviour()
}

\description{
  This function is aimed primarily at developers extending
  this package. You shouldn't need to use it for everyday
  operation of the package.

  Roxygen defaults to using all subclasses of
  \code{\link{Tag}}, and all processors
  (\code{$process[A-Z].*}) and writers
  (\code{$writer[A-Z].*}) defined in the roxygen package.
}

\details{

}

% Generated by roxygen from [behaviour.r:8:14]
