\name{default_behaviour}

\alias{default_behaviour}

\title{[DEV] Encapuslate default behaviour.}

\usage{
  default_behaviour()
}

\description{
  This function is useful only for developers

  Roxygen defaults to using all subclasses of
  \code{\link{Tag}}, and all processors
  (\code{$process[A-Z].*}) and writers
  (\code{$writer[A-Z].*}) defined in the roxygen package.
}

\details{

}

