% Built by roxygen3 0.1
\name{defaultTag}
\alias{defaultTag}
\title{[DEV] Generate default tag.}
\usage{
  defaultTag(tag, object)
}
\arguments{
  \item{tag}{class of tag}

  \item{object}{\linkS4class{Object} to generate defaults for.}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

Generate default tag.
}
\section{Method signatures for generic defaultTag}{
\itemize{
  \item \linkS4class{AliasesTag} (S4), \linkS4class{PackageObject} (S4)
  \item \linkS4class{DebugTag} (S4), \linkS4class{Object} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{FunctionObject} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{R5ClassObject} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{S3GenericObject} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{S3MethodObject} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{S4ClassObject} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{S4GenericObject} (S4)
  \item \linkS4class{DefaultExportTag} (S4), \linkS4class{S4MethodObject} (S4)
  \item \linkS4class{DocTypeTag} (S4), \linkS4class{DataObject} (S4)
  \item \linkS4class{DocTypeTag} (S4), \linkS4class{PackageObject} (S4)
  \item \linkS4class{DocTypeTag} (S4), \linkS4class{R5ClassObject} (S4)
  \item \linkS4class{DocTypeTag} (S4), \linkS4class{S4ClassObject} (S4)
  \item \linkS4class{DocTypeTag} (S4), \linkS4class{S4MethodObject} (S4)
  \item \linkS4class{FormatTag} (S4), \linkS4class{DataObject} (S4)
  \item \linkS4class{KeywordsTag} (S4), \linkS4class{DataObject} (S4)
  \item \linkS4class{KeywordsTag} (S4), \linkS4class{S3GenericObject} (S4)
  \item \linkS4class{KeywordsTag} (S4), \linkS4class{S3MethodObject} (S4)
  \item \linkS4class{KeywordsTag} (S4), \linkS4class{S4ClassObject} (S4)
  \item \linkS4class{KeywordsTag} (S4), \linkS4class{S4GenericObject} (S4)
  \item \linkS4class{KeywordsTag} (S4), \linkS4class{S4MethodObject} (S4)
  \item \linkS4class{NameTag} (S4), \linkS4class{NullObject} (S4)
  \item \linkS4class{NameTag} (S4), \linkS4class{Object} (S4)
  \item \linkS4class{NameTag} (S4), \linkS4class{S4ClassObject} (S4)
  \item \linkS4class{NameTag} (S4), \linkS4class{S4MethodObject} (S4)
  \item \linkS4class{UsageTag} (S4), \linkS4class{DataObject} (S4)
  \item \linkS4class{UsageTag} (S4), \linkS4class{FunctionObject} (S4)
  \item \linkS4class{UsageTag} (S4), \linkS4class{S3MethodObject} (S4)
  \item \linkS4class{UsageTag} (S4), \linkS4class{S4GenericObject} (S4)
  \item \linkS4class{UsageTag} (S4), \linkS4class{S4MethodObject} (S4)
}

}

\keyword{methods}
