% Built by roxygen3 0.1
\name{defaultTag}
\alias{defaultTag}
\title{[DEV] Generate default tag.}
\usage{
  defaultTag(tag, object)
}
\arguments{
  \item{tag}{class of tag}

  \item{object}{\linkS4class{Object} to generate defaults for.}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

Generate default tag.
}
\section{Method signatures for generic defaultTag}{
\itemize{
  \item \linkS4class{AliasesTag}, \linkS4class{PackageObject}
  \item \linkS4class{DebugTag}, \linkS4class{Object}
  \item \linkS4class{DefaultExportTag}, \linkS4class{FunctionObject}
  \item \linkS4class{DefaultExportTag}, \linkS4class{R5ClassObject}
  \item \linkS4class{DefaultExportTag}, \linkS4class{S3GenericObject}
  \item \linkS4class{DefaultExportTag}, \linkS4class{S3MethodObject}
  \item \linkS4class{DefaultExportTag}, \linkS4class{S4ClassObject}
  \item \linkS4class{DefaultExportTag}, \linkS4class{S4GenericObject}
  \item \linkS4class{DefaultExportTag}, \linkS4class{S4MethodObject}
  \item \linkS4class{DocTypeTag}, \linkS4class{DataObject}
  \item \linkS4class{DocTypeTag}, \linkS4class{PackageObject}
  \item \linkS4class{DocTypeTag}, \linkS4class{R5ClassObject}
  \item \linkS4class{DocTypeTag}, \linkS4class{S4ClassObject}
  \item \linkS4class{DocTypeTag}, \linkS4class{S4MethodObject}
  \item \linkS4class{FormatTag}, \linkS4class{DataObject}
  \item \linkS4class{KeywordsTag}, \linkS4class{DataObject}
  \item \linkS4class{KeywordsTag}, \linkS4class{S3GenericObject}
  \item \linkS4class{KeywordsTag}, \linkS4class{S3MethodObject}
  \item \linkS4class{KeywordsTag}, \linkS4class{S4ClassObject}
  \item \linkS4class{KeywordsTag}, \linkS4class{S4GenericObject}
  \item \linkS4class{KeywordsTag}, \linkS4class{S4MethodObject}
  \item \linkS4class{NameTag}, \linkS4class{NullObject}
  \item \linkS4class{NameTag}, \linkS4class{Object}
  \item \linkS4class{NameTag}, \linkS4class{S4ClassObject}
  \item \linkS4class{NameTag}, \linkS4class{S4MethodObject}
  \item \linkS4class{UsageTag}, \linkS4class{DataObject}
  \item \linkS4class{UsageTag}, \linkS4class{FunctionObject}
  \item \linkS4class{UsageTag}, \linkS4class{S3MethodObject}
  \item \linkS4class{UsageTag}, \linkS4class{S4GenericObject}
  \item \linkS4class{UsageTag}, \linkS4class{S4MethodObject}
}

}

\keyword{methods}
