% Built by roxygen3 0.1
\docType{class}
\name{DescriptionTag-class}
\alias{DescriptionTag-class}
\title{Override default description.}
\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\linkS4class{TagIntro}} for more details.
}
\section{Tag Usage}{
  \code{ #' @description Text goes here. }
}

