\docType{class}
\name{DescriptionTag-class}
\alias{DescriptionTag-class}
\title{Override default description.}
\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\link{tag_intro}} for more details.
}
\section{Tag Usage}{
  \code{ #' @description Text goes here. }
}

% Generated by roxygen from [tag-description.r:7:7]
