% Built by roxygen3 0.1
\docType{class}
\name{DetailsTag-class}
\alias{DetailsTag-class}
\title{@details: provide (lots of) extra detail about the function}
\description{
  The topic title. By default this is taken from the third
  and subsequent paragraphs of the roxygen block. See
  \code{\linkS4class{IntroTag}} for more details.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @details Text goes here.
}
}
