% Built by roxygen3 0.1
\docType{class}
\name{DocTypeTag-class}
\alias{DocTypeTag-class}
\title{Set object documentation type.}
\description{
  Set object documentation type.
}
\details{
  You can use any doctype, but it will only be included in
  the Rd file if it is one of the standard R doctypes:
  data, package, methods and class.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @docType data
#'  @docType package
#'  @docType custom doctype
}
}
