% Built by roxygen3 0.1
\docType{class}
\name{NameTag-class}
\alias{NameTag-class}
\title{@name: Override the default topic name.}
\description{
By default, the topic name is derived from the object following the
roxygen comments. This tag is rarely needed and using it may suggest
that you are documenting the wrong object.
}
\seealso{
@rdname for modifying the generated file name and combining
  the documentation for multiple objects into one file.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @name name
}
}
