% Built by roxygen3 0.1
\docType{class}
\name{Object-class}
\alias{DataObject-class}
\alias{FunctionObject-class}
\alias{NullObject-class}
\alias{Object-class}
\alias{PackageObject-class}
\alias{R5ClassObject-class}
\alias{S3GenericObject-class}
\alias{S3MethodObject-class}
\alias{S4ClassObject-class}
\alias{S4GenericObject-class}
\alias{S4MethodObject-class}
\title{[DEV] The object being documented.}
\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

The object being documented.
}
\section{Class hierarchy for Object}{
Sub classes:
\itemize{
  \item \linkS4class{NullObject}
  \item \linkS4class{PackageObject}
  \item \linkS4class{DataObject}
  \item \linkS4class{FunctionObject}
  \item \linkS4class{S4ClassObject}
  \item \linkS4class{S4MethodObject}
  \item \linkS4class{S4GenericObject}
  \item \linkS4class{S3MethodObject}
  \item \linkS4class{S3GenericObject}
  \item \linkS4class{R5ClassObject}
}

}

\section{Generics with methods for Object}{
\itemize{
  \item \code{\link{defaultTag}}* (30 methods defined for subclasses)
  \item \code{\link{isNull}}* (1 methods defined for subclasses)
  \item \code{\link{location}} (1 methods defined for subclasses)
  \item \code{\link{show}}
}

* = methods only defined for subclasses
}

\keyword{classes}
