\name{object_from_call}

\alias{object_from_call}

\title{Given a call that modifies the R environment, find the object that
it creates}

\usage{
  object_from_call(call, env)
}

\arguments{
  \item{call}{unevaluated function call}

  \item{env}{environment in which to evaluate function
  call}
}

\value{
  a list giving the \code{name} and \code{value} of the
  object that the call creates.  \code{NULL} is returned if
  the call doesn't modify the package environment in a way
  that roxygen recognises.
}

\description{
  Given a call that modifies the R environment, find the
  object that it creates.
}

\examples{
a <- 1
object_from_call(quote(a <- 1), environment())
}

\keyword{internal}

\comment{Generated by roxygen from [R/object-from-call.r]}

