\docType{class}
\name{PackageBundle-class}

\alias{PackageBundle}
\alias{PackageBundle-class}

\title{PackageBundle class.}

\usage{
  PackageBundle(path, behaviour = default_behaviour())
}

\description{
  The package class captures all the information about the
  package: its name, path, and all the \code{\link{Block}}s
  that it contains.
}

\details{

}

% Generated by roxygen from [all-classes.r:71:75]

% Generated by roxygen from [bundle-package.r:7:20]
