\name{parse_block}

\alias{parse_block}

\title{Parse and execute a block of text in a package like environment.}

\usage{
  parse_block(text, tags = base_tags())
}

\arguments{
  \item{text}{code to parse/execute}
}

\description{
  This is used cheifly for testing.
}

\details{

}

% Generated by roxygen from [parse.r:50:64]
