% Built by roxygen3 0.1
\name{parse_block}
\alias{parse_block}
\title{Parse and execute a block of text in a package like environment.}
\usage{
  parse_block(text, behaviour = no_output())
}
\arguments{
  \item{text}{code to parse/execute}

  \item{behaviour}{a \code{\linkS4class{Behaviour}} object, specifying which
  tags, processors and writers to use.}
}

\description{
This is used cheifly for testing.
}
\keyword{internal}
