\name{parse_directory}
\alias{parse_directory}
\title{Parse directory of source files.}
\usage{
  parse_directory(path, env, tags = base_tags())
}
\arguments{
  \item{path}{path to directory of R files to parse}

  \item{env}{environment that contains the results of
  evaluating the file.}
}

\description{
  Parse directory of source files.
}
\keyword{internal}
% Generated by roxygen from [parse.r:6:11]
