\name{parse_file}

\alias{parse_file}

\title{[DEV]}

\arguments{
  \item{path}{path of file to parse}
}

\value{
  A list of roc objects
}

\description{
  This function is useful only for developers
}

\details{
  Parse a source file containing roxygen blocks.
}

\comment{Generated by roxygen from [R/rocblock-read.r]}

