\name{parse_file}

\alias{parse_file}

\title{[DEV] Parse a source file containing roxygen blocks.}

\usage{
  parse_file(path, env = NULL)
}

\arguments{
  \item{path}{path of file to parse}

  \item{env}{environment that contains the results of
  evaluating the file.  If \code{NULL}, will create a new
  environment with the global environment as a parent and
  will evaluate the code in that environment. This will be
  fine for simple, stand alone files, but you'll need to
  manage the environments yourself for more complicated
  packages.}
}

\value{
  A list of roc objects
}

\description{
  This function is useful only for developers

  Parse a source file containing roxygen blocks.
}

% Generated by roxygen from [R/rocblock-parse.r]
