\name{parse_file}

\alias{parse_file}

\title{[DEV] Parse a source file containing roxygen blocks.}

\usage{
  parse_file(path, env = NULL, tags = base_tags())
}

\arguments{
  \item{path}{path of file to parse}
}

\value{
  A list of roc objects
}

\description{
  This function is useful only for developers

  Parse a source file containing roxygen blocks.
}

\details{

}

