\name{parse_file}

\alias{parse_file}

\title{Parse a source file containing roxygen blocks.}

\usage{
  parse_file(path, env, tags = base_tags())
}

\arguments{
  \item{path}{path of file to parse}

  \item{env}{environment that contains the results of
  evaluating the file.}
}

\value{
  A list of roc objects
}

\description{
  Parse a source file containing roxygen blocks.
}

\keyword{internal}

% Generated by roxygen from [parse.r:19:25]
