\docType{methods}
\name{procBlock}

\alias{}
\alias{procBlock}

\title{Title, description and details.}

\usage{
  \S4method{procBlock}{TagDev}(tag, block)

  \S4method{procBlock}{TagExample}(tag, block)

  \S4method{procBlock}{TagExport}(tag, block)

  \S4method{procBlock}{TagName}(tag, block)

  \S4method{procBlock}{TagS3method}(tag, block)

  \S4method{procBlock}{TagAutoImports}(tag, block)

  \S4method{procBlock}{TagPath}(tag, block)

  \S4method{procBlock}{Tag}(tag, block)
}

\description{
  Title, description and details.
}

\details{

}

\seealso{
  \code{\link{tag_title}}, \code{\link{tag_description}},
  \code{\link{tag_details}} to set each component
  individually.
}

