\docType{methods}
\name{procBlock}

\alias{}
\alias{procBlock}

\title{Title, description and details.}

\usage{
  \S4method{procBlock}{TagDev}(tag, block)

  \S4method{procBlock}{TagExample}(tag, block)

  \S4method{procBlock}{TagExport}(tag, block)

  \S4method{procBlock}{TagName}(tag, block)

  \S4method{procBlock}{TagS3method}(tag, block)

  \S4method{procBlock}{TagAutoImports}(tag, block)

  \S4method{procBlock}{Tag}(tag, block)
}

\description{
  Title, description and details.
}

\details{

}

\seealso{
  \code{\link{tag_title}}, \code{\link{tag_description}},
  \code{\link{tag_details}} to set each component
  individually.
}

% Generated by roxygen from [roxy-process.r:33:35]

% Generated by roxygen from [tag-dev.r:3:14]

% Generated by roxygen from [tag-example.r:2:8]

% Generated by roxygen from [tag-export.r:27:33]

% Generated by roxygen from [tag-intro.r:15:48]

% Generated by roxygen from [tag-name.r:11:14]

% Generated by roxygen from [tag-namespace-export.r:42:61]

% Generated by roxygen from [tag-namespace-import.r:44:60]

% Generated by roxygen from [tag.r:17:17]
