% Built by roxygen3 0.1
\name{process}
\alias{process}
\title{[DEV] Process bundles and blocks.}
\usage{
  process(input, ...)
}
\arguments{
  \item{input}{object to process}

  \item{...}{other arguments used by methods.}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

Process bundles and blocks.
}
\section{Method signatures for generic process}{
\itemize{
  \item \linkS4class{AutoImportsTag}
  \item \linkS4class{AutoSlotsTag}
  \item \linkS4class{Block}
  \item \linkS4class{Bundle}
  \item \linkS4class{ClassHierarchyTag}
  \item \linkS4class{ClassMethodsTag}
  \item \linkS4class{DevTag}
  \item \linkS4class{ExampleTag}
  \item \linkS4class{ExportTag}
  \item \linkS4class{GenericMethodsTag}
  \item \linkS4class{IntroTag}
  \item \linkS4class{NameTag}
  \item \linkS4class{S3methodTag}
  \item \linkS4class{SlotTag}
  \item \linkS4class{Tag}
  \item \linkS4class{UseDynLibTag}
}

}

\keyword{methods}
