% Built by roxygen3 0.1
\name{process}
\alias{process}
\title{[DEV] Process bundles and blocks.}
\usage{
  process(input, ...)
}
\arguments{
  \item{input}{object to process}

  \item{...}{other arguments used by methods.}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

Process bundles and blocks.
}
\section{Method signatures for generic process}{
\itemize{
  \item \linkS4class{AutoImportsTag} (S4)
  \item \linkS4class{AutoSlotsTag} (S4)
  \item \linkS4class{Block} (S4)
  \item \linkS4class{Bundle} (S4)
  \item \linkS4class{ClassHierarchyTag} (S4)
  \item \linkS4class{ClassMethodsTag} (S4)
  \item \linkS4class{DevTag} (S4)
  \item \linkS4class{ExampleTag} (S4)
  \item \linkS4class{ExportTag} (S4)
  \item \linkS4class{GenericMethodsTag} (S4)
  \item \linkS4class{IntroTag} (S4)
  \item \linkS4class{NameTag} (S4)
  \item \linkS4class{S3methodTag} (S4)
  \item \linkS4class{SlotTag} (S4)
  \item \linkS4class{Tag} (S4)
  \item \linkS4class{UseDynLibTag} (S4)
}

}

\keyword{methods}
