\name{rd_out}

\alias{rd_out}

\title{[DEV]}

\usage{
  rd_out(tag, name = NULL)
}

\arguments{
  \item{tag}{a function that takes a single argument (tag)
  as input, and returns a list of \code{rd_commands} as
  output}

  \item{name}{name of the input tag - this is usually not
  specified as it is filled in by \code{\link{roccer}} but
  is useful for testing.}
}

\description{
  This function is useful only for developers
}

\details{
  Output to Rd files in the man directory.

  This uses the \code{@rdname} tag to determine which file
  the output from each rocblock is sent to.

  Only one of \code{tag} and \code{out} can be supplied.
}

\comment{Generated by roxygen from [R/output-rd.r]}

