\docType{class}
\name{RoxyPackage}

\alias{RoxyPackage}

\title{RoxyPackage class.}

\description{
  The package class captures all the information about the
  package: its name, path, and all the
  \code{\link{RoxyBlock}}s that it contains.
}

\details{

}

% Generated by roxygen from [roxy-package.r:19:20]
