% Built by roxygen3 0.1
\name{rPath}
\alias{rPath}
\title{Compute path to a bundle.}
\usage{
  rPath(bundle)
}
\arguments{
  \item{bundle}{input bundle}
}

\description{
Compute path to a bundle.
}
\section{Method signatures for generic rPath}{
\itemize{
  \item \linkS4class{Bundle} (S4)
  \item \linkS4class{DirectoryBundle} (S4)
  \item \linkS4class{PackageBundle} (S4)
}

}

\keyword{internal}
