\docType{roccer}

\name{@_intro}

\alias{@_intro}

\title{\code{\link{}

\usage{
  #' The first line becomes the topic title.

  The next paragraph goes into the descrption section.  Any
    of these sections can span multiple lines.

  The third and subsequent paragraph go into the details.
}

\description{
  \code{\link{
}

\details{
  Title, description and details.

  to set each component individually.
}

\seealso{
  \code{\link{
}

\comment{Generated by roxygen from [R/roccer-intro.r]}

