\docType{roccer}

\name{@description}

\alias{@description}

\title{Override default description.}

\usage{
  #' @description Text goes here.
}

\description{
  The topic title. By default this is taken from the second
  paragraph of the roxygen block. See
  \code{\link{roc_intro}} for more details.
}

\comment{Generated by roxygen from [R/roccer-title-desc-details.r]}

