\docType{roccer}

\name{@details}

\alias{@details}

\title{Override default details.}

\usage{
  #' @details Text goes here.
}

\description{
  The topic title. By default this is taken from the third
  and subsequent paragraphs of the roxygen block. See
  \code{\link{roc_intro}} for more details.
}

\comment{Generated by roxygen from [R/roccer-title-desc-details.r]}

