\docType{class}
\name{ExportTag-class}

\alias{ExportClassTag-class}
\alias{ExportMethodsTag-class}
\alias{ExportPatternTag-class}
\alias{ExportTag-class}
\alias{S3methodTag-class}

\title{Namespace: tags for exporting objects}

\usage{
  @export @export function name

  @S3method generic class @S3method generic @S3method
}

\description{
  Generally, you will only need to use a bare
  \code{@export} - this will inspect the documented object
  and automatically figure out the correct export (which
  varies based on whether it's a function, s3 method, s4
  class, s4 generic/method).
}

\details{
  The convention in \pkg{roxygen3} is that exporting a
  generic automatically exports all associated methods.
  This happens automatically for S4 with the
  \code{exportMethod} directive, but needs considerable
  work for S3. For details, see \link{roxgyen_s3}.

  It's not recommend practice to use \code{@exportPattern}
  (instead it's better to individually label the functions
  to be exported), but this may be useful for legacy
  packages.

  \code{@S3method} exists largely for compatibility with
  roxygen2. Roxygen3 now automatically determines if an
  object is an S3 method, and so only \code{@export is
  necessary.}
}

\section{Tag Usage}{
  \code{ #' @exportClass class1 class2 }
}

\section{Tag Usage}{
  \code{ #' @exportMethods generic1 generic2 }
}

\section{Tag Usage}{
  \code{ #' @exportPattern pattern }
}

% Generated by roxygen from [tag-export.r:25:25]

% Generated by roxygen from [tag-namespace-export.r:4:4]

% Generated by roxygen from [tag-namespace-export.r:14:14]

% Generated by roxygen from [tag-namespace-export.r:24:24]

% Generated by roxygen from [tag-namespace-export.r:37:38]
