\docType{class}
\name{TagImportFrom}

\alias{TagAutoImports}
\alias{TagImport}
\alias{TagImportClassesFrom}
\alias{TagImportFrom}
\alias{TagImportMethodsFrom}

\title{Namespace: tags for importing functions.}

\usage{
  @importFrom package function1 function2

  @autoImports

  @import package1 package2 package3

  @importClassesFrom package fun1 fun2

  @importMethodsFrom package fun1 fun2
}

\description{
  By and large, \code{@autoImports} should be the only
  imports tag that you need to use. It automatically
  generates the necessary \code{importFrom} statements to
  import all external functions used by this function.  See
  \code{\link{auto_imports}} for more implementation
  details.
}

\details{
  If there is a conflict, use \code{tag_importFrom} to
  resolve it. You can do \code{@importFrom base function} -
  this is not normally allowed in the \file{NAMESPACE}, but
  roxygen3 will simply ignore it, but still use it when
  resolving conflicts.

  You must have the packages declared in \code{DESCRIPTION}
  Imports.
}

% Generated by roxygen from [tag-namespace-import.r:19:20]

% Generated by roxygen from [tag-namespace-import.r:44:44]

% Generated by roxygen from [tag-namespace-import.r:68:68]

% Generated by roxygen from [tag-namespace-import.r:79:79]

% Generated by roxygen from [tag-namespace-import.r:90:90]
