
\name{tag_importFrom}

\alias{tag_auto_import}
\alias{tag_import}
\alias{tag_importClassesFrom}
\alias{tag_importFrom}
\alias{tag_importMethodsFrom}

\title{Namespace: tags for importing functions.}

\usage{
  #' @importFrom package function1 function2

  #' @auto_import

  #' @import package1 package2 package3

  #' @importClassesFrom package fun1 fun2

  #' @importMethodsFrom package fun1 fun2
}

\description{
  By and large, \code{@auto_imports} should be the only
  imports tag that you need to use. It automatically
  generates the necessary \code{importFrom} statements to
  import all external functions used by this function.  See
  \code{\link{auto_imports}} for more implementation
  details.
}

\details{
  If there is a conflict, use \code{tag_importFrom} to
  resolve it. You can do \code{@importFrom base function} -
  this is not normally allowed in the \file{NAMESPACE}, but
  roxygen3 will simply ignore it, but still use it when
  resolving conflicts.

  You must have the packages declared in \code{DESCRIPTION}
  Imports.
}

% Generated by roxygen from [R/roccer-namespace-import.r]

% Generated by roxygen from [R/roccer-namespace-import.r]

% Generated by roxygen from [R/roccer-namespace-import.r]

% Generated by roxygen from [R/roccer-namespace-import.r]

% Generated by roxygen from [R/roccer-namespace-import.r]
