\docType{roccer}

\name{@inheritParams}

\alias{@inheritParams}

\title{\item{\code{@inheritParams source_function}}{Alternatively, you can
   inherit parameter description from another function. This tag will
   bring in all documentation for parameters that are undocumented in the
   current function, but documented in the source function. The source 
   can be a function in the current package, \code{function}, or another
   package \code{package::function}.}}

\description{
  \item{\code{@inheritParams
  source_function}}{Alternatively, you can inherit
  parameter description from another function. This tag
  will bring in all documentation for parameters that are
  undocumented in the current function, but documented in
  the source function. The source can be a function in the
  current package, \code{function}, or another package
  \code{package::function}.}
}

% Generated by roxygen from [R/roccer-inherit-params.r]
