
\name{tag_title}

\alias{tag_title}

\title{Override default topic title.}

\usage{
  #' @title Topic title
}

\description{
  The topic title. By default this is taken from the first
  paragraph of the roxygen block. See
  \code{\link{tag__intro}} for more details.
}

% Generated by roxygen from [R/roccer-title-desc-details.r]
