\docType{class}
\name{Tag-class}

\alias{Tag}
\alias{Tag-class}

\title{Tag class}

\usage{
  Tag(text)
}

\description{
  The tag class is the base class for all roxygen3 tags.
}

\section{Parsing order}{
  The functions \code{procTag} and \code{process} are
  called in that order, so if both are supplied,
  \code{process} can rely on the tag already being
  parsed.
}

% Generated by roxygen from [all-classes.r:7:15]

% Generated by roxygen from [tag.r:9:11]
