% Built by roxygen3 0.1
\docType{class}
\name{Tag-class}
\alias{Tag}
\alias{Tag-class}
\title{[DEV] Tag class}
\usage{
  Tag()
}
\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

The tag class is the base class for all roxygen3 tags.  To implement a new
tag, you will need to subclass this class.
}
\section{Class hierarchy for Tag}{
Sub classes:
\itemize{
  \item \linkS4class{AliasesTag}
  \item \linkS4class{AuthorTag}
  \item \linkS4class{AutoImportsTag}
  \item \linkS4class{ClassHierarchyTag}
  \item \linkS4class{ClassMethodsTag}
  \item \linkS4class{CollateTag}
  \item \linkS4class{CommentTag}
  \item \linkS4class{ConceptTag}
  \item \linkS4class{DebugTag}
  \item \linkS4class{DefaultExportTag}
  \item \linkS4class{DescriptionTag}
  \item \linkS4class{DetailsTag}
  \item \linkS4class{DevTag}
  \item \linkS4class{DocTypeTag}
  \item \linkS4class{EncodingTag}
  \item \linkS4class{ExamplesTag}
  \item \linkS4class{ExampleTag}
  \item \linkS4class{ExportClassTag}
  \item \linkS4class{ExportMethodsTag}
  \item \linkS4class{ExportPatternTag}
  \item \linkS4class{ExportTag}
  \item \linkS4class{FamilyTag}
  \item \linkS4class{FormatTag}
  \item \linkS4class{GenericMethodsTag}
  \item \linkS4class{ImportClassesFromTag}
  \item \linkS4class{ImportFromTag}
  \item \linkS4class{ImportMethodsFromTag}
  \item \linkS4class{ImportTag}
  \item \linkS4class{IncludeTag}
  \item \linkS4class{InheritParamsTag}
  \item \linkS4class{IntroTag}
  \item \linkS4class{KeywordsTag}
  \item \linkS4class{MethodTag}
  \item \linkS4class{NameTag}
  \item \linkS4class{NoRdTag}
  \item \linkS4class{NoteTag}
  \item \linkS4class{ParamTag}
  \item \linkS4class{RdnameTag}
  \item \linkS4class{ReferencesTag}
  \item \linkS4class{ReturnTag}
  \item \linkS4class{S3methodTag}
  \item \linkS4class{SectionTag}
  \item \linkS4class{SeealsoTag}
  \item \linkS4class{SlotTag}
  \item \linkS4class{AutoSlotsTag}
  \item \linkS4class{SourceTag}
  \item \linkS4class{TitleTag}
  \item \linkS4class{TagUsageTag}
  \item \linkS4class{UsageTag}
  \item \linkS4class{UseDynLibTag}
}

}

\section{Generics with methods for Tag}{
\itemize{
  \item \code{\link[=format]{format}}
  \item \code{\link[=isEmpty]{isEmpty}}
  \item \code{\link[=process]{process}}
  \item \code{\link[=show]{show}}
  \item \code{\link[=value]{value}}
  \item \code{\link[=value<-]{value<-}}
  \item \code{\link[=writeDescription]{writeDescription}}
  \item \code{\link[=writeNamespace]{writeNamespace}}
  \item \code{\link[=writeRd]{writeRd}}
}

}

\keyword{classes}
