\docType{class}
\name{TagAliases}

\alias{TagAliases}

\title{Add additional topic aliases.}

\usage{
  @aliases space separated aliases
}

\description{
  Add additional aliases through which the user can find
  the documentation with \code{\link{help}}. The topic name
  is always included in the list of aliases.
}

\details{

}

% Generated by roxygen from [tag-aliases.r:9:9]
