\docType{class}
\name{TagDocType}

\alias{TagDocType}

\title{Set object documentation type.}

\usage{
  @docType data @docType package @docType custom doctype
}

\description{
  Set object documentation type.
}

\details{
  You can use any doctype, but it will only be included in
  the Rd file if it is one of the standard R doctypes:
  data, package, methods and class.
}

% Generated by roxygen from [tag-doctype.r:17:17]
