\docType{class}
\name{TagDocType}

\alias{TagDocType}

\title{Set object documentation type.}

\description{
  Set object documentation type.
}

\details{
  You can use any doctype, but it will only be included in
  the Rd file if it is one of the standard R doctypes:
  data, package, methods and class.
}

\section{Tag Usage}{
  \code{ #' @docType data @docType package @docType custom
  doctype }
}

% Generated by roxygen from [tag-doctype.r:12:12]
