% Built by roxygen3 0.1
\docType{class}
\name{TitleTag-class}
\alias{TitleTag-class}
\title{Override default topic title.}
\description{
  The topic title. By default this is taken from the first
  paragraph of the roxygen block. See
  \code{\linkS4class{TagIntro}} for more details.
}
\section{Tag Usage}{
  \code{ #' @title Topic title }
}

