% Built by roxygen3 0.1
\docType{class}
\name{TitleTag-class}
\alias{TitleTag-class}
\title{@title: Override default topic title.}
\description{
  The topic title. By default this is taken from the first
  paragraph of the roxygen block. See
  \code{\linkS4class{IntroTag}} for more details.
}
\keyword{classes}
\section{Tag Usage}{
\preformatted{
#' @title Topic title
}
}
