\name{usage}

\alias{usage}

\title{Generate usage statement for an R object.}

\usage{
  usage(obj, name)
}

\arguments{
  \item{obj}{object to generate usage for}

  \item{name}{name of object (if not computable from
  \code{obj})}
}

\description{
  Generate usage statement for an R object.
}

\comment{Generated by roxygen from [R/usage.r]}

