% Built by roxygen3 0.1
\name{writeNamespace}
\alias{writeNamespace}
\title{[DEV] An output generator for the \file{NAMESPACE} file.}
\usage{
  writeNamespace(object)
}
\arguments{
  \item{object}{Object to proccess, starting at a \linkS4class{Bundle},
  breaking down into \linkS4class{Block}s then individual
  \linkS4class{Tag}s}
}

\description{
This function is aimed primarily at developers extending this package.
You shouldn't need to use it for everyday operation.

An output generator for the \file{NAMESPACE} file.
}
\keyword{methods}
