\docType{methods}
\name{writeRd}

\alias{writeRd}

\title{[DEV] Output to Rd files in the man directory.}

\usage{
  writeRd(object)

  \S4method{writeRd}{TagAliases}(object)

  \S4method{writeRd}{TagAuthor}(object)

  \S4method{writeRd}{TagComment}(object)

  \S4method{writeRd}{TagDescription}(object)

  \S4method{writeRd}{TagDetails}(object)

  \S4method{writeRd}{TagDocType}(object)

  \S4method{writeRd}{TagExamples}(object)

  \S4method{writeRd}{TagFormat}(object)

  \S4method{writeRd}{TagKeywords}(object)

  \S4method{writeRd}{TagName}(object)

  \S4method{writeRd}{TagNote}(object)

  \S4method{writeRd}{TagParam}(object)

  \S4method{writeRd}{TagPath}(object)

  \S4method{writeRd}{TagReferences}(object)

  \S4method{writeRd}{TagReturn}(object)

  \S4method{writeRd}{TagSection}(object)

  \S4method{writeRd}{TagSeealso}(object)

  \S4method{writeRd}{TagSeealso}(object)

  \S4method{writeRd}{TagSource}(object)

  \S4method{writeRd}{TagTagUsage}(object)

  \S4method{writeRd}{TagTitle}(object)

  \S4method{writeRd}{TagUsage}(object)

  \S4method{writeRd}{RoxyPackage}(object)

  \S4method{writeRd}{RoxyBundle}(object)

  \S4method{writeRd}{RoxyBlock}(object)

  \S4method{writeRd}{Tag}(object)
}

\arguments{
  \item{tag}{a function that takes a single argument (tag)
  as input, and returns a list of \code{rd_commands} as
  output}

  \item{name}{input tag name, usually set by
  \code{\link{roccer}}.}
}

\description{
  This function is useful only for developers

  This uses the \code{@rdname} tag to determine which file
  the output from each rocblock is sent to.
}

\details{
  Only one of \code{tag} and \code{out} can be supplied.
}

% Generated by roxygen from [all-generics.r:5:7]

% Generated by roxygen from [tag-aliases.r:15:17]

% Generated by roxygen from [tag-author.r:2:4]

% Generated by roxygen from [tag-comments.r:7:9]

% Generated by roxygen from [tag-description.r:8:10]

% Generated by roxygen from [tag-details.r:9:11]

% Generated by roxygen from [tag-doctype.r:40:42]

% Generated by roxygen from [tag-examples.r:2:4]

% Generated by roxygen from [tag-format.r:11:13]

% Generated by roxygen from [tag-keywords.r:11:13]

% Generated by roxygen from [tag-name.r:22:24]

% Generated by roxygen from [tag-note.r:2:4]

% Generated by roxygen from [tag-param.r:17:19]

% Generated by roxygen from [tag-path.r:12:16]

% Generated by roxygen from [tag-references.r:2:4]

% Generated by roxygen from [tag-return.r:3:5]

% Generated by roxygen from [tag-section.r:14:16]

% Generated by roxygen from [tag-seealso.r:2:4]

% Generated by roxygen from [tag-seelaso.r:2:4]

% Generated by roxygen from [tag-source.r:2:4]

% Generated by roxygen from [tag-tag-usage.r:3:6]

% Generated by roxygen from [tag-title.r:8:10]

% Generated by roxygen from [tag-usage.r:14:16]

% Generated by roxygen from [write-rd.r:13:15]

% Generated by roxygen from [write-rd.r:16:19]

% Generated by roxygen from [write-rd.r:20:22]

% Generated by roxygen from [write-rd.r:23:23]
